improvement: refactory code
