+++ src/constants/constants.js
@@ -1,6 +1,3 @@
-import * as yup from 'yup';
-import moment from 'moment';
-
 const IdBoard = import.meta.env.VITE_ID_BOARD;
 const IdListPedAberto = `624a047f2f06001532cef5e5`;
 
@@ -23,27 +20,6 @@ const idCustomFields = {
   idCustomPagamento: '62845d1714150e5f18da882e',
 };
 
-const validationScheme = {
-  nameInOrder: yup
-    .string()
-    .required('Campo obrigatório')
-    .min(10, 'Por favor, escreva seu nome completo'),
-  celInOrder: yup.string().required('Campo obrigatório'),
-  phraseOnTheCake: yup
-    .string()
-    .required('Campo obrigatório')
-    .max(52, 'A quantidade de caracteres excede o espaço no bolo'),
-  cakePhraseColor: yup.string().required('Campo obrigatório'),
-  cakeColor: yup.string().required('Campo obrigatório'),
-  flavorInOrder: yup.string().required('Campo obrigatório').nullable(),
-  isWithdrawal: yup.string().required('Campo obrigatório').nullable(),
-  candleInOrder: yup.string().required('Campo obrigatório').nullable(),
-  formOfPaymentInOrder: yup.string().required('Campo obrigatório').nullable(),
-  dateTimeInOrder: yup
-    .date()
-    .min(moment().add(10, 'm').toDate(), 'Selecione um horário válido de ')
-    .required('Campo obrigatório'),
-};
 
 export {
   IdBoard,
@@ -53,7 +29,6 @@ export {
   urlTrelloGetCustomField,
   urlTrelloPostCard,
   idCustomFields,
-  validationScheme,
   Labels,
   Entrega,
   Retirada,
